+++ framework/domain/src/main/java/is/codion/framework/domain/entity/attribute/ValueAttributeDefinition.java
@@ -124,7 +124,7 @@ public sealed interface ValueAttributeDefinition<T> extends AttributeDefinition<
 	 * @param entity the entity containing the value to validate
 	 * @param nullable true if null values are allowed in this validation context,
 	 * false if null should trigger a null validation exception
-	 * @throws ValidationException in case of one or more invalid values
+	 * @throws ValidationException in case the value of this attribute is invalid in the given entity
 	 * @see EntityDefinition.Builder#validator(EntityValidator)
 	 * @see Builder#validator(AttributeValidator)
 	 * @see AttributeValidator